Create storm.py
Loads in pickle file with burstiness.py calculations for the entire cosmological timescale. Core slopes are loaded in from a txt file with a value corresponding to each halo in the sim. Burstiness values are also calculated during active star formation only and instantaneously during a constrained timescale near the end of the simulation. The mean, active, or instantaneous burstiness values and core slopes are plotted as function of the stellar-mass fraction which is calculated for each halo in the MARVEL-ous Storm blastwave and superbubble feedback sims. 

A linear regression model is included to evaluate the slope, R^2, and y-intercept for any burstiness and core slope value set.